Adding Json payload
I need to add a Json payload, because a service that I use requires that
the object being sent be a string and with the header
content-type:application/json set. I had to create a new command-line
argument and a new payload reader, to read the contents of a file with a
json into a string.